+++ packages/maas/frontend/src/app/pages/api-keys/utils.ts
@@ -1,5 +1,3 @@
-import { TokenRateLimitInfo, UserSubscription } from '../../types/subscriptions';
-
 export enum ModelSource {
   Internal = 'Internal',
   External = 'External',
@@ -24,63 +22,6 @@ export const formatApiKeyError = (message: string): string => {
     return `Requested expiration exceeds maximum allowed (${maxExpirationMatch[1]} days). Select a shorter duration and try again.`;
   }
   return message.charAt(0).toUpperCase() + message.slice(1);
-};
-
-export type ModelGroupSubscription = {
-  subscriptionIdHeader: string;
-  displayName?: string;
-  keyCount?: number;
-  tokenRateLimits?: TokenRateLimitInfo[];
-};
-
-export type ModelGroupEntry = {
-  name: string;
-  displayName?: string;
-  description?: string;
-  source?: string;
-  subscriptions: ModelGroupSubscription[];
-};
-
-export const formatTokenLimit = (limits?: TokenRateLimitInfo[]): string => {
-  if (!limits || limits.length === 0) {
-    return '—';
-  }
-  return limits
-    .map((l) => {
-      const formatted = l.limit >= 1000 ? `${Math.floor(l.limit / 1000)}K` : String(l.limit);
-      return `${formatted} / ${l.window}`;
-    })
-    .join(', ');
-};
-
-export const deriveModelGroups = (subscriptions: UserSubscription[]): ModelGroupEntry[] => {
-  const modelMap = new Map<string, ModelGroupEntry>();
-
-  subscriptions.forEach((sub) => {
-    sub.model_refs.forEach((ref) => {
-      const existing = modelMap.get(ref.name);
-      const subEntry: ModelGroupSubscription = {
-        subscriptionIdHeader: sub.subscription_id_header,
-        displayName: sub.display_name,
-        keyCount: sub.key_count,
-        tokenRateLimits: ref.token_rate_limits,
-      };
-
-      if (existing) {
-        existing.subscriptions.push(subEntry);
-      } else {
-        modelMap.set(ref.name, {
-          name: ref.name,
-          displayName: ref.display_name,
-          description: ref.description,
-          source: ref.source,
-          subscriptions: [subEntry],
-        });
-      }
-    });
-  });
-
-  return Array.from(modelMap.values());
 };
 
 export const getSourceLabelColor = (source: string): 'blue' | 'purple' =>
